added first implementations for binary search
The speedup you can observe for the interleaved functions is only partially
coming from utilized memory parallelism. The compiler produces branchless code
for some branches which results in a significant speedup as well. To really
compare apples with apples, we have to force branchless code with inline
assembly and not rely on the compiler to maybe do it.